Reversed the AH49's power switch, slowed down to 5 min cadence

+++ lp-gauge/src/deep_sleep.hpp
@@ -10,7 +10,7 @@ ISR (WDT_vect)
 }
 
 
-void deep_sleep(unsigned msec)
+void deep_sleep(unsigned long msec)
 {
 #ifdef DISABLE_SLEEP
     delay(msec);
